Ui class: Add ui add to handle user interface message TaskManager class: Add taskManger to handle the function of the type XanChatBot class: Update the class to make it more OOP
