+++ src/main/data/xan.txt
@@ -1,5 +1,3 @@
 [E][X] project meeting (from: Mon 2pm to: 4pm)
 [T][ ] borrow book
-[T][ ] read book
-[T][ ] visit new theme park
-[E][ ] project meeting (from: Mon 2pm to: 4pm)
+[T][X] visit new theme park
